Update field representation to comply with scratch update
Scratch updated the way it represents fixed input (drop-down) fields.
E.g. the fields for the ftduino led block, where the block can be "On" and "Off".
Previously the ftduino_led block had a ftduino_menu_ONOFFSTATE block as input.
This now the ftudino_led block has a *field* that represents an on off state.